Introduce bitmap_value_detail.h
