+++ be/src/types/bitmap_value.h
@@ -1,6 +1,6 @@
 // This file is made available under Elastic License 2.0.
 // This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/be/src/types/bitmap_value.h
+//   https://github.com/apache/incubator-doris/blob/master/be/src/util/bitmap_value.h
 
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
